Move FakeFormController to its own file
